chore: change IIFE to class
